Remove generated query client stub changes
